# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /bookings/{bookingId}/check-in:
    post:
      operationId: SafetyController_checkIn
      parameters:
      - name: bookingId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInDto'
      responses:
        '200':
          description: Check-in created
      summary: Check in to a trek
      tags:
      - Safety
  /bookings/{bookingId}/check-in-status:
    get:
      operationId: SafetyController_getCheckInStatus
      parameters:
      - name: bookingId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Check-in status
      summary: Get check-in/out status
      tags:
      - Safety
  /bookings/{bookingId}/check-out:
    post:
      operationId: SafetyController_checkOut
      parameters:
      - name: bookingId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckOutDto'
      responses:
        '200':
          description: Check-out completed
      summary: Check out from a trek
      tags:
      - Safety
...
  /check-in/{checkInId}/acknowledge:
    post:
      operationId: SafetyController_acknowledge
      parameters:
      - name: checkInId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeSafetyDto'
      responses:
        '200':
          description: Safety acknowledged
      summary: Acknowledge safety and cancel escalation
      tags:
      - Safety
...
  /organizer/treks/{id}:
    get:
      operationId: OrganizerController_getTrekDetail
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: Get organizer trek detail
      tags:
      - Organizer
...
  /profile/emergency-contacts:
    get:
      operationId: SafetyController_getEmergencyContacts
      parameters: []
      responses:
        '200':
          description: List of emergency contacts
      summary: List user emergency contacts
      tags:
      - Safety
    post:
      operationId: SafetyController_addEmergencyContact
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmergencyContactDto'
      responses:
        '200':
          description: Emergency contact created
      summary: Add emergency contact
      tags:
      - Safety
  /profile/emergency-contacts/{id}:
    patch:
      operationId: SafetyController_updateEmergencyContact
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmergencyContactDto'
      responses:
        '200':
          description: Emergency contact updated
      summary: Update emergency contact
      tags:
      - Safety
    delete:
      operationId: SafetyController_deleteEmergencyContact
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Emergency contact deleted
      summary: Delete emergency contact
      tags:
      - Safety
  /recommendations:
    get:
      operationId: RecommendationsController_getRecommendations
      parameters:
      - name: limit
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecommendationResultDto'
      summary: Get personalized trek recommendations
      tags:
      - Recommendations
...
  /treks:
    post:
      operationId: TreksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrekDto'
      responses:
        '200':
          description: Created trek
      summary: Create a trek
      tags:
      - Treks
    get:
      operationId: TreksController_list
      parameters:
      - name: q
        required: false
        in: query
        schema:
          type: string
      - name: state
        required: false
        in: query
        schema:
          type: string
      - name: difficulty
        required: false
        in: query
        schema:
          type: string
          enum:
          - EASY
          - MODERATE
          - DIFFICULT
          - EXTREME
      - name: minCost
        required: false
        in: query
        schema:
          minimum: 0
          type: number
      - name: maxCost
        required: false
        in: query
        schema:
          minimum: 0
          type: number
      - name: sort
        required: false
        in: query
        schema:
          type: string
          enum:
          - relevance
          - newest
          - popular
          - distance
      - name: page
        required: false
        in: query
        schema:
          type: number
      - name: limit
        required: false
        in: query
        schema:
          type: number
      responses:
        '200':
          description: Paginated trek list
      summary: List and search treks
      tags:
      - Treks
  /treks/nearby:
    get:
      operationId: TreksController_nearby
      parameters:
      - name: latitude
        required: true
        in: query
        schema:
          type: number
      - name: longitude
        required: true
        in: query
        schema:
          type: number
      - name: radiusMeters
        required: false
        in: query
        schema:
          minimum: 10
          default: 5000
          type: number
      - name: page
        required: false
        in: query
        schema:
          type: number
      - name: limit
        required: false
        in: query
        schema:
          type: number
      responses:
        '200':
          description: Treks near the provided coordinates
      summary: Find treks near a location
      tags:
      - Treks
  /treks/recommendations:
    get:
      operationId: TreksController_recommendations
      parameters:
      - name: limit
        required: false
        in: query
        schema:
          type: number
      - name: lat
        required: false
        in: query
        schema:
          type: number
      - name: lon
        required: false
        in: query
        schema:
          type: number
      responses:
        '200':
          description: Recommended treks
      summary: Get trek recommendations
      tags:
      - Treks
  /treks/{id}:
    get:
      operationId: TreksController_get
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Trek details
      summary: Get trek details
      tags:
      - Treks
...
  /treks/{trekId}/recommendations:
    get:
      operationId: RecommendationsController_getSimilarTreks
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      - name: limit
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarTrekDto'
      summary: Get similar treks for a specific trek
      tags:
      - Recommendations
  /treks/{trekId}/safety:
    get:
      operationId: SafetyController_getTrekSafety
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Safety information for the trek
      summary: Get safety info for a trek
      tags:
      - Safety
    put:
      operationId: SafetyController_upsertTrekSafety
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSafetyInfoDto'
      responses:
        '200':
          description: Safety info upserted
      summary: Upsert safety info for a trek
      tags:
      - Safety
...
info:
  title: Offbeat प्रवासी API
  description: Full-scale, production-ready trek discovery and booking platform.
  version: 1.0.0
  contact: {}
...
servers:
- url: http://localhost:4000
  description: Local development
- url: https://api.offbeatpravasi.com
  description: Production
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    AcknowledgeSafetyDto:
      type: object
      properties:
        message:
          type: string
...
    CheckInDto:
      type: object
      properties:
        latitude:
          type: number
          description: Latitude of check-in location
        longitude:
          type: number
          description: Longitude of check-in location
      required:
      - latitude
      - longitude
    CheckOutDto:
      type: object
      properties:
        latitude:
          type: number
          description: Latitude of check-out location
        longitude:
          type: number
          description: Longitude of check-out location
...
    CreateEmergencyContactDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 20
        relationship:
          type: string
          maxLength: 40
        isPrimary:
          type: boolean
      required:
      - name
      - phone
      - relationship
...
    CreateTrekDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        shortDescription:
          type: string
          maxLength: 512
        fullDescription:
          type: string
        state:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        difficulty:
          type: string
          enum:
          - EASY
          - MODERATE
          - DIFFICULT
          - EXTREME
        costInr:
          type: number
          minimum: 0
        maxParticipants:
          type: number
          minimum: 1
        tags:
          type: array
          items:
            type: string
        imageKeys:
          type: array
          items:
            type: string
      required:
      - name
...
    RecommendationResultDto:
      type: object
      properties:
        trekId:
          type: string
        score:
          type: number
        reason:
          type: string
      required:
      - trekId
      - score
      - reason
...
    SimilarTrekDto:
      type: object
      properties:
        trekId:
          type: string
        score:
          type: number
        reason:
          type: string
      required:
      - trekId
      - score
      - reason
...
    UpdateEmergencyContactDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 20
        relationship:
          type: string
          maxLength: 40
        isPrimary:
          type: boolean
...
    UpsertSafetyInfoDto:
      type: object
      properties:
        terrainRisks:
          type: string
        altitudeWarnings:
          type: string
        wildlifeAdvisories:
          type: string
        generalGuidelines:
          type: string
        baseCampContact:
          type: string
          maxLength: 32
        localRescueContact:
          type: string
          maxLength: 32
        nearestHospital:
          type: string
          maxLength: 255
...
security:
- ApiKey: []